requirements.txt
# Discord Bot
discord.py>=2.3.0

# Docker Integration
docker>=6.0.0

# Environment Variables
python-dotenv>=1.0.0

# Data Handling
pydantic>=2.0.0

# Logging
colorlog>=6.7.0

# Testing (Development)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Code Quality (Development)
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation (Development)
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
